cello_delegate microbenchmark
