+++ apps/cello_delegate_test/.gitignore
@@ -1,0 +1,1 @@
+*/
