+++ apps/cello_delegate_test/app_path.mk
@@ -1,0 +1,1 @@
+APP_PATH = $(HB_HAMMERBENCH_PATH)/apps/cello_delegate_path
