+++ apps/cello_delegate_test/tests.mk
@@ -1,0 +1,7 @@
+####################
+# CHANGE ME: TESTS #
+####################
+# TESTS += $(call test-name,[n])
+TESTS += $(call test-name,1)
+
+
